31385 muon data plot refactor step 1
