+++ scripts/Muon/GUI/Common/plot_widget/plot_muon_data_pane_model.py
@@ -1,0 +1,79 @@
+# Mantid Repository : https://github.com/mantidproject/mantid
+#
+# Copyright &copy; 2020 ISIS Rutherford Appleton Laboratory UKRI,
+#   NScD Oak Ridge National Laboratory, European Spallation Source,
+#   Institut Laue - Langevin & CSNS, Institute of High Energy Physics, CAS
+# SPDX - License - Identifier: GPL - 3.0 +
+#from Muon.GUI.Common.utilities.run_string_utils import run_list_to_string
+
+
+class PlotMuonDataPanetModel(object):
+
+    def __init__(self, context):
+        self.context = context
+
+    def get_workspace_list_to_plot(self, is_raw, plot_type):
+        """
+         :return: a list of workspace names to plot
+         """
+        workspace_list = self.get_workspaces_to_plot(is_raw, plot_type)
+        return workspace_list
+
+    def get_workspace_list_and_indices_to_plot(self, is_raw, plot_type):
+        """
+         :return: a list of workspace names to plot
+         """
+        workspace_list = self.get_workspaces_to_plot(is_raw, plot_type)
+        indices = self._generate_run_indices(workspace_list)
+
+        return workspace_list, indices
+
+    def get_workspaces_to_plot(self, is_raw, plot_type):
+        """
+        :param is_raw: Whether to use raw or rebinned data
+        :param plot_type: plotting type, e.g Counts, Frequency Re
+        :return: a list of workspace names
+        """
+        currently_selected = self.context.group_pair_context.selected_groups_and_pairs
+        workspace_list = []
+        for grouppair in currently_selected:
+            workspace_list += self.get_time_workspaces_to_plot(grouppair, is_raw, plot_type)
+        return workspace_list
+
+    def get_time_workspaces_to_plot(self, current_group_pair, is_raw, plot_type):
+        """
+        :param current_group_pair: The group/pair currently selected
+        :param is_raw: Whether to use raw or rebinned data
+        :param plot_type: Whether to plot counts or asymmetry
+        :return: a list of workspace names
+        """
+        try:
+            if is_raw:
+                workspace_list = self.context.group_pair_context[current_group_pair].get_asymmetry_workspace_names(
+                    self.context.data_context.current_runs)
+            else:
+                workspace_list = self.context.group_pair_context[
+                    current_group_pair].get_asymmetry_workspace_names_rebinned(
+                    self.context.data_context.current_runs)
+
+            if plot_type == "Counts":
+                workspace_list = [item.replace("Asymmetry", "Counts")
+                                  for item in workspace_list if "Asymmetry" in item]
+
+            return workspace_list
+        except AttributeError:
+            return []
+
+    @staticmethod
+    def _generate_run_indices(workspace_list):
+        indices = [0 for _ in workspace_list]
+        return indices
+
+    def get_workspace_and_indices_for_group_or_pair(self, group_or_pair:str, is_raw, plot_type):
+        """
+         :return: a list of workspace names and corresponding indices to plot
+         """
+        workspace_list = self.get_time_workspaces_to_plot(group_or_pair, is_raw, plot_type)
+        indices = self._generate_run_indices(workspace_list)
+
+        return workspace_list, indices
