+++ scripts/Muon/GUI/Common/plot_widget/plotting_widget_view.ui
@@ -23,20 +23,6 @@
       <number>0</number>
      </property>
      <item>
-      <widget class="QComboBox" name="fit_or_data_combo_box">
-       <item>
-        <property name="text">
-         <string>Plot Data</string>
-        </property>
-       </item>
-       <item>
-        <property name="text">
-         <string>Plot Fits</string>
-        </property>
-       </item>
-      </widget>
-     </item>
-     <item>
       <widget class="QComboBox" name="plot_type_combo"/>
      </item>
      <item>
